(core): new types based on type classes
the original api was not very consistent in the types approach,
as it was still relying on inheritance and the separation
of the actual graph types and the value types did not really work
out for implementing the cats instances.

most points should be fixed by that, it also includes a more complete
mxGraph integration.